+++ addOns/network/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Fixed
 - Do not pass-through requests to the local proxies themselves (e.g. ZAP domain, aliases).
+- Correctly handle concurrent requests (Issue 7838).
 
 ## [0.7.0] - 2023-04-04
 ### Changed